feat: exported IconType from ui and migrated characterCounter to Input from forms/FormFieldInput
